requirement.txt
Django==4.2.7
djangorestframework==3.14.0
transformers==4.30.0
torch==2.1.0
numpy==1.24.2
requests==2.31.0
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.0.0
python-dotenv==1.0.0
pypdf==3.8.1  # Added pypdf
